+++ src/Nest/ElasticClient-Snapshot.cs
@@ -1,9 +1,4 @@
 ﻿using System;
-using System.CodeDom;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using Elasticsearch.Net;
 
@@ -193,130 +188,4 @@ namespace Nest
 			);
 		}
 	}
-
-    public class SnapshotObservable : IDisposable, IObservable<ISnapshotStatusResponse>
-    {
-        private readonly IElasticClient _elasticClient;
-        private readonly ISnapshotRequest _snapshotRequest;
-        private readonly TimeSpan _interval = TimeSpan.FromSeconds(2);
-        private Timer _timer;
-        private bool _disposed;
-
-        public SnapshotObservable(IElasticClient elasticClient, ISnapshotRequest snapshotRequest)
-        {
-            _elasticClient = elasticClient;
-            _snapshotRequest = snapshotRequest;
-        }
-
-        public SnapshotObservable(IElasticClient elasticClient, ISnapshotRequest snapshotRequest, TimeSpan interval)
-            :this(elasticClient, snapshotRequest)
-        {
-            interval.ThrowIfNull("interval");
-            if (interval.Ticks < 0) throw new ArgumentOutOfRangeException("interval");
-
-            _interval = interval;
-        }
-
-        public IDisposable Subscribe(IObserver<ISnapshotStatusResponse> observer)
-        {
-            observer.ThrowIfNull("observer");
-
-            try
-            {
-                _snapshotRequest.RequestParameters.WaitForCompletion(false);
-                this._elasticClient.Snapshot(_snapshotRequest);
-
-                _timer = new Timer(Snapshot, observer, _interval.Milliseconds, Timeout.Infinite);
-            }
-            catch (Exception exception)
-            {
-                observer.OnError(exception);
-            }
-
-            return this;
-        }
-
-        private void Snapshot(object state)
-        {
-            var observer = state as IObserver<ISnapshotStatusResponse>;
-
-            if(observer == null) throw new ArgumentException("state");
-
-            try
-            {
-                var watch = new Stopwatch();
-                watch.Start();
-
-                var snapshotStatusResponse = this._elasticClient.SnapshotStatus(descriptor => descriptor
-                    .Repository(_snapshotRequest.Repository)
-                    .Snapshot(_snapshotRequest.Snapshot));
-
-                if (!snapshotStatusResponse.IsValid)
-                    throw new SnapshotException(snapshotStatusResponse.ConnectionStatus, "Can't create snapshot");
-
-                if (snapshotStatusResponse.Snapshots.All(s => s.ShardsStats.Done == s.ShardsStats.Total))
-                {
-                    observer.OnCompleted();
-                    _timer.Change(Timeout.Infinite, Timeout.Infinite);
-                    return;
-                }
-
-                observer.OnNext(snapshotStatusResponse);
-                _timer.Change(Math.Max(0, _interval.Milliseconds - watch.ElapsedMilliseconds), Timeout.Infinite);
-            }
-            catch (Exception exception)
-            {
-                observer.OnError(exception);
-                _timer.Change(Timeout.Infinite, Timeout.Infinite);
-            }
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (_disposed) return;
-
-            _timer.Dispose();
-
-            _disposed = true;
-        }
-
-        ~SnapshotObservable()
-        {
-            Dispose(false);
-        }
-    }
-
-    public class Observer<T> : IObserver<T>
-    {
-        private readonly Action<T> _onNext;
-        private readonly Action<Exception> _onError;
-        private readonly Action _completed;
-
-        public Observer(Action<T> onNext = null, Action<Exception> onError = null, Action completed = null)
-        {
-            _onNext = onNext;
-            _onError = onError;
-            _completed = completed;
-        }
-
-        public void OnNext(T value)
-        {
-            if (this._onNext != null) this._onNext(value);
-        }
-
-        public void OnError(Exception error)
-        {
-            if (this._onError != null) this._onError(error);
-        }
-
-        public void OnCompleted()
-        {
-            if (this._completed != null) this._completed();
-        }
-    }
 }
