+++ src/Nest/Nest.csproj
@@ -197,6 +197,9 @@
     <Compile Include="Domain\Responses\ExplainResponse.cs" />
     <Compile Include="Domain\Responses\GetFieldMappingResponse.cs" />
     <Compile Include="Domain\Cat\ICatRecord.cs" />
+    <Compile Include="Domain\Responses\Observer.cs" />
+    <Compile Include="Domain\Responses\RestoreObservable.cs" />
+    <Compile Include="Domain\Responses\SnapshotObservable.cs" />
     <Compile Include="Domain\Responses\UpgradeResponse.cs" />
     <Compile Include="Domain\Responses\UpgradeStatusResponse.cs" />
     <Compile Include="Domain\Responses\VerifyRepositoryResponse.cs" />
